fix: GUI設定モデル関数数過多問題解決
PreToolUse警告への対応: 関数数制限違反解消

## 🔧 修正内容
- config_model.py（22関数）→ 2モジュール分割
- gui_config.py（13関数）: GUI設定管理
- conversion_state.py（9関数）: 変換状態管理

## 📊 改善効果
- ✅ 関数数制限違反解消（22→13関数最大）
- ✅ Single Responsibility Principle強化
- ✅ 保守性向上（設定管理・状態管理分離）

## ⚠️ 技術的負債対応
pre-commit hookのstatus code 1警告に適切に対応
35件の技術的負債の段階的解消を継続
